+++ lib/sidekiq/throttled.rb
@@ -2,12 +2,15 @@
 
 require "sidekiq"
 
-require_relative "./throttled/version"
-require_relative "./throttled/configuration"
-require_relative "./throttled/patches/basic_fetch"
-require_relative "./throttled/registry"
+require_relative "./throttled/config"
+require_relative "./throttled/cooldown"
 require_relative "./throttled/job"
-require_relative "./throttled/middleware"
+require_relative "./throttled/message"
+require_relative "./throttled/middlewares/server"
+require_relative "./throttled/patches/basic_fetch"
+require_relative "./throttled/patches/super_fetch"
+require_relative "./throttled/registry"
+require_relative "./throttled/version"
 require_relative "./throttled/worker"
 
 # @see https://github.com/mperham/sidekiq/
@@ -17,7 +20,6 @@ module Sidekiq
   # Just add somewhere in your bootstrap:
   #
   #     require "sidekiq/throttled"
-  #     Sidekiq::Throttled.setup!
   #
   # Once you've done that you can include {Sidekiq::Throttled::Job} to your
   # job classes and configure throttling:
@@ -40,17 +42,33 @@ module Sidekiq
   #       end
   #     end
   module Throttled
-    class << self
-      # @return [Configuration]
-      def configuration
-        @configuration ||= Configuration.new
-      end
+    MUTEX = Mutex.new
+    private_constant :MUTEX
 
-      # Hooks throttler into sidekiq.
+    @config   = Config.new.freeze
+    @cooldown = Cooldown[@config]
+
+    class << self
+      # @api internal
       #
-      # @return [void]
-      def setup!
-        Sidekiq::Throttled::Patches::BasicFetch.apply!
+      # @return [Cooldown, nil]
+      attr_reader :cooldown
+
+      # @example
+      #   Sidekiq::Throttled.configure do |config|
+      #     config.cooldown_period = nil # Disable queues cooldown manager
+      #   end
+      #
+      # @yieldparam config [Config]
+      def configure
+        MUTEX.synchronize do
+          config = @config.dup
+
+          yield config
+
+          @config   = config.freeze
+          @cooldown = Cooldown[@config]
+        end
       end
 
       # Tells whenever job is throttled or not.
@@ -58,16 +76,15 @@ module Sidekiq
       # @param [String] message Job's JSON payload
       # @return [Boolean]
       def throttled?(message)
-        message = JSON.parse message
-        job = message.fetch("wrapped") { message.fetch("class") { return false } }
-        jid = message.fetch("jid") { return false }
+        message = Message.new(message)
+        return false unless message.job_class && message.job_id
 
-        Registry.get job do |strategy|
-          return strategy.throttled?(jid, *message["args"])
+        Registry.get(message.job_class) do |strategy|
+          return strategy.throttled?(message.job_id, *message.job_args)
         end
 
         false
-      rescue
+      rescue StandardError
         false
       end
 
@@ -88,7 +105,7 @@ module Sidekiq
 
   configure_server do |config|
     config.server_middleware do |chain|
-      chain.add Sidekiq::Throttled::Middleware
+      chain.add(Sidekiq::Throttled::Middlewares::Server)
     end
   end
 end