+++ spec/lib/sidekiq/throttled/registry_spec.rb
@@ -59,7 +59,7 @@ RSpec.describe Sidekiq::Throttled::Registry do
     let(:name) { "foo" }
 
     context "when strategy is not registered" do
-      it { is_expected.to be_nil }
+      it { is_expected.to be(nil) }
     end
 
     context "when strategy was registered" do
@@ -76,15 +76,7 @@ RSpec.describe Sidekiq::Throttled::Registry do
 
       before { described_class.add(parent_class.name, **threshold) }
 
-      it { is_expected.to be_nil }
-
-      context "when configuration has inherit strategy turned on" do
-        before { Sidekiq::Throttled.configuration.inherit_strategies = true }
-
-        after { Sidekiq::Throttled.configuration.reset! }
-
-        it { is_expected.to be described_class.get("Parent") }
-      end
+      it { is_expected.to be described_class.get("Parent") }
     end
   end